Safari-609.1.11.1/LayoutTests/accessibility/aria-modal-expected.txt

This tests that aria-modal on dialog makes other elements inert.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS backgroundAccessible() is false
PASS backgroundAccessible() is true
PASS backgroundAccessible() is false
PASS okBtn.isIgnored is false
PASS backgroundAccessible() is true
PASS backgroundAccessible() is true
PASS backgroundAccessible() is false
PASS backgroundAccessible() is true
PASS successfullyParsed is true

TEST COMPLETE
Other page content with a dummy focusable element

Display a dialog
